Package name changes for classes which moved from kafka->stream
