+++ pinot-integration-tests/src/test/java/org/apache/pinot/integration/tests/ClusterTest.java
@@ -62,8 +62,8 @@ import org.apache.pinot.controller.helix.ControllerRequestBuilderUtil;
 import org.apache.pinot.controller.helix.ControllerTest;
 import org.apache.pinot.core.data.GenericRow;
 import org.apache.pinot.core.indexsegment.generator.SegmentVersion;
-import org.apache.pinot.core.realtime.impl.kafka.AvroRecordToPinotRowGenerator;
 import org.apache.pinot.core.realtime.impl.kafka.KafkaStreamConfigProperties;
+import org.apache.pinot.core.realtime.stream.AvroRecordToPinotRowGenerator;
 import org.apache.pinot.core.realtime.stream.StreamConfig;
 import org.apache.pinot.core.realtime.stream.StreamConfigProperties;
 import org.apache.pinot.core.realtime.stream.StreamMessageDecoder;
